refractor into many mods, halfway into get_configuration
