+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 actix-web = "4.3.1"
+config = "0.13.3"
 reqwest = "0.11.18"
 serde = { version = "1.0.163", features = ["derive"] }
 tokio = "1.28.2"
