+++ src/lib.rs
@@ -1,35 +1,3 @@
-use actix_web::dev::Server;
-use actix_web::{get, web, App, HttpResponse, HttpServer, Responder};
-use serde::Deserialize;
-use std::net::TcpListener;
-
-#[get("/")]
-async fn hello() -> impl Responder {
-    HttpResponse::Ok().body("Hello world!")
-}
-
-async fn health_check() -> impl Responder {
-    HttpResponse::Ok()
-}
-
-#[derive(serde::Deserialize)]
-struct SubscribeFormData {
-    email: String,
-    name: String,
-}
-
-async fn subscribe(_form: web::Form<SubscribeFormData>) -> impl Responder {
-    HttpResponse::Ok()
-}
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .service(hello)
-            .route("/subscriptions", web::post().to(subscribe))
-            .route("/health_check", web::get().to(health_check))
-    })
-    .listen(listener)?
-    .run();
-    Ok(server)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
