+++ src/main.rs
@@ -1,9 +1,8 @@
-use zero2prod::run;
 use std::net::TcpListener;
+use zero2prod::startup::run;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    let listener = TcpListener::bind("127.0.0.1:8000")
-        .expect("Failed to bind to port.");
+    let listener = TcpListener::bind("127.0.0.1:8000").expect("Failed to bind to port.");
     run(listener)?.await
 }
