+++ src/routes/hello.rs
@@ -1,0 +1,6 @@
+use actix_web::{get, HttpResponse, Responder};
+
+#[get("/")]
+async fn hello() -> impl Responder {
+    HttpResponse::Ok().body("Hello world!")
+}
