+++ src/routes/mod.rs
@@ -1,0 +1,7 @@
+mod health_check;
+mod hello;
+mod subscriptions;
+
+pub use health_check::*;
+pub use hello::*;
+pub use subscriptions::*;
