+++ src/routes/subscriptions.rs
@@ -1,0 +1,12 @@
+use actix_web::{web, HttpResponse, Responder};
+use serde::Deserialize;
+
+#[derive(serde::Deserialize)]
+pub struct SubscribeFormData {
+    email: String,
+    name: String,
+}
+
+pub async fn subscribe(_form: web::Form<SubscribeFormData>) -> impl Responder {
+    HttpResponse::Ok()
+}
